Split UI code into multiple files.
