+++ phoneblock_answerbot_ui/lib/proto.dart
@@ -48,17 +48,11 @@ abstract class _JsonObject {
 }
 
 /// Visitor interface for SetupRequest.
-abstract class SetupRequestVisitor<R, A> {
+abstract class SetupRequestVisitor<R, A> implements BotRequestVisitor<R, A> {
 	R visitCreateAnswerBot(CreateAnswerBot self, A arg);
-	R visitEnterHostName(EnterHostName self, A arg);
-	R visitSetupDynDns(SetupDynDns self, A arg);
-	R visitCheckDynDns(CheckDynDns self, A arg);
-	R visitEnableAnswerBot(EnableAnswerBot self, A arg);
-	R visitDisableAnswerBot(DisableAnswerBot self, A arg);
-	R visitDeleteAnswerBot(DeleteAnswerBot self, A arg);
-	R visitCheckAnswerBot(CheckAnswerBot self, A arg);
 }
 
+///  Base class for all supported requests.
 abstract class SetupRequest extends _JsonObject {
 	/// Creates a SetupRequest.
 	SetupRequest();
@@ -86,6 +80,7 @@ abstract class SetupRequest extends _JsonObject {
 			case "DisableAnswerBot": result = DisableAnswerBot(); break;
 			case "DeleteAnswerBot": result = DeleteAnswerBot(); break;
 			case "CheckAnswerBot": result = CheckAnswerBot(); break;
+			case "ListCalls": result = ListCalls(); break;
 			default: result = null;
 		}
 
@@ -195,14 +190,101 @@ class CreateAnswerBot extends SetupRequest {
 
 }
 
-class EnterHostName extends SetupRequest {
+/// Visitor interface for BotRequest.
+abstract class BotRequestVisitor<R, A> {
+	R visitEnterHostName(EnterHostName self, A arg);
+	R visitSetupDynDns(SetupDynDns self, A arg);
+	R visitCheckDynDns(CheckDynDns self, A arg);
+	R visitEnableAnswerBot(EnableAnswerBot self, A arg);
+	R visitDisableAnswerBot(DisableAnswerBot self, A arg);
+	R visitDeleteAnswerBot(DeleteAnswerBot self, A arg);
+	R visitCheckAnswerBot(CheckAnswerBot self, A arg);
+	R visitListCalls(ListCalls self, A arg);
+}
+
+///  Base class for all requests targeting a single answer bot.
+abstract class BotRequest extends SetupRequest {
+	///  The ID of the answer bot this request is targeted to
 	int id;
 
+	/// Creates a BotRequest.
+	BotRequest({
+			this.id = 0, 
+	});
+
+	/// Parses a BotRequest from a string source.
+	static BotRequest? fromString(String source) {
+		return read(JsonReader.fromString(source));
+	}
+
+	/// Reads a BotRequest instance from the given reader.
+	static BotRequest? read(JsonReader json) {
+		BotRequest? result;
+
+		json.expectArray();
+		if (!json.hasNext()) {
+			return null;
+		}
+
+		switch (json.expectString()) {
+			case "EnterHostName": result = EnterHostName(); break;
+			case "SetupDynDns": result = SetupDynDns(); break;
+			case "CheckDynDns": result = CheckDynDns(); break;
+			case "EnableAnswerBot": result = EnableAnswerBot(); break;
+			case "DisableAnswerBot": result = DisableAnswerBot(); break;
+			case "DeleteAnswerBot": result = DeleteAnswerBot(); break;
+			case "CheckAnswerBot": result = CheckAnswerBot(); break;
+			case "ListCalls": result = ListCalls(); break;
+			default: result = null;
+		}
+
+		if (!json.hasNext() || json.tryNull()) {
+			return null;
+		}
+
+		if (result == null) {
+			json.skipAnyValue();
+		} else {
+			result._readContent(json);
+		}
+		json.endArray();
+
+		return result;
+	}
+
+	@override
+	void _readProperty(String key, JsonReader json) {
+		switch (key) {
+			case "id": {
+				id = json.expectInt();
+				break;
+			}
+			default: super._readProperty(key, json);
+		}
+	}
+
+	@override
+	void _writeProperties(JsonSink json) {
+		super._writeProperties(json);
+
+		json.addKey("id");
+		json.addNumber(id);
+	}
+
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg);
+
+	@override
+	R visitSetupRequest<R, A>(SetupRequestVisitor<R, A> v, A arg) => visitBotRequest(v, arg);
+
+}
+
+///  Sets the host name of the Fritz!Box to connect the answer bot to.
+class EnterHostName extends BotRequest {
 	String hostName;
 
 	/// Creates a EnterHostName.
 	EnterHostName({
-			this.id = 0, 
+			super.id, 
 			this.hostName = "", 
 	});
 
@@ -224,10 +306,6 @@ class EnterHostName extends SetupRequest {
 	@override
 	void _readProperty(String key, JsonReader json) {
 		switch (key) {
-			case "id": {
-				id = json.expectInt();
-				break;
-			}
 			case "hostName": {
 				hostName = json.expectString();
 				break;
@@ -240,26 +318,22 @@ class EnterHostName extends SetupRequest {
 	void _writeProperties(JsonSink json) {
 		super._writeProperties(json);
 
-		json.addKey("id");
-		json.addNumber(id);
-
 		json.addKey("hostName");
 		json.addString(hostName);
 	}
 
 	@override
-	R visitSetupRequest<R, A>(SetupRequestVisitor<R, A> v, A arg) => v.visitEnterHostName(this, arg);
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg) => v.visitEnterHostName(this, arg);
 
 }
 
-class SetupDynDns extends SetupRequest {
-	int id;
-
+///  Sets up a DnyDNS account for the Fritz!Box to register its IP address.
+class SetupDynDns extends BotRequest {
 	String hostName;
 
 	/// Creates a SetupDynDns.
 	SetupDynDns({
-			this.id = 0, 
+			super.id, 
 			this.hostName = "", 
 	});
 
@@ -281,10 +355,6 @@ class SetupDynDns extends SetupRequest {
 	@override
 	void _readProperty(String key, JsonReader json) {
 		switch (key) {
-			case "id": {
-				id = json.expectInt();
-				break;
-			}
 			case "hostName": {
 				hostName = json.expectString();
 				break;
@@ -297,15 +367,12 @@ class SetupDynDns extends SetupRequest {
 	void _writeProperties(JsonSink json) {
 		super._writeProperties(json);
 
-		json.addKey("id");
-		json.addNumber(id);
-
 		json.addKey("hostName");
 		json.addString(hostName);
 	}
 
 	@override
-	R visitSetupRequest<R, A>(SetupRequestVisitor<R, A> v, A arg) => v.visitSetupDynDns(this, arg);
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg) => v.visitSetupDynDns(this, arg);
 
 }
 
@@ -373,12 +440,11 @@ class SetupDynDnsResponse extends _JsonObject {
 
 }
 
-class CheckDynDns extends SetupRequest {
-	int id;
-
+///  Checks, whether a DynDNS request has been received.
+class CheckDynDns extends BotRequest {
 	/// Creates a CheckDynDns.
 	CheckDynDns({
-			this.id = 0, 
+			super.id, 
 	});
 
 	/// Parses a CheckDynDns from a string source.
@@ -397,35 +463,15 @@ class CheckDynDns extends SetupRequest {
 	String _jsonType() => "CheckDynDns";
 
 	@override
-	void _readProperty(String key, JsonReader json) {
-		switch (key) {
-			case "id": {
-				id = json.expectInt();
-				break;
-			}
-			default: super._readProperty(key, json);
-		}
-	}
-
-	@override
-	void _writeProperties(JsonSink json) {
-		super._writeProperties(json);
-
-		json.addKey("id");
-		json.addNumber(id);
-	}
-
-	@override
-	R visitSetupRequest<R, A>(SetupRequestVisitor<R, A> v, A arg) => v.visitCheckDynDns(this, arg);
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg) => v.visitCheckDynDns(this, arg);
 
 }
 
-class EnableAnswerBot extends SetupRequest {
-	int id;
-
+///  Switches the answer bot on.
+class EnableAnswerBot extends BotRequest {
 	/// Creates a EnableAnswerBot.
 	EnableAnswerBot({
-			this.id = 0, 
+			super.id, 
 	});
 
 	/// Parses a EnableAnswerBot from a string source.
@@ -444,35 +490,15 @@ class EnableAnswerBot extends SetupRequest {
 	String _jsonType() => "EnableAnswerBot";
 
 	@override
-	void _readProperty(String key, JsonReader json) {
-		switch (key) {
-			case "id": {
-				id = json.expectInt();
-				break;
-			}
-			default: super._readProperty(key, json);
-		}
-	}
-
-	@override
-	void _writeProperties(JsonSink json) {
-		super._writeProperties(json);
-
-		json.addKey("id");
-		json.addNumber(id);
-	}
-
-	@override
-	R visitSetupRequest<R, A>(SetupRequestVisitor<R, A> v, A arg) => v.visitEnableAnswerBot(this, arg);
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg) => v.visitEnableAnswerBot(this, arg);
 
 }
 
-class DisableAnswerBot extends SetupRequest {
-	int id;
-
+///  Switches the answer bot off.
+class DisableAnswerBot extends BotRequest {
 	/// Creates a DisableAnswerBot.
 	DisableAnswerBot({
-			this.id = 0, 
+			super.id, 
 	});
 
 	/// Parses a DisableAnswerBot from a string source.
@@ -491,35 +517,15 @@ class DisableAnswerBot extends SetupRequest {
 	String _jsonType() => "DisableAnswerBot";
 
 	@override
-	void _readProperty(String key, JsonReader json) {
-		switch (key) {
-			case "id": {
-				id = json.expectInt();
-				break;
-			}
-			default: super._readProperty(key, json);
-		}
-	}
-
-	@override
-	void _writeProperties(JsonSink json) {
-		super._writeProperties(json);
-
-		json.addKey("id");
-		json.addNumber(id);
-	}
-
-	@override
-	R visitSetupRequest<R, A>(SetupRequestVisitor<R, A> v, A arg) => v.visitDisableAnswerBot(this, arg);
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg) => v.visitDisableAnswerBot(this, arg);
 
 }
 
-class DeleteAnswerBot extends SetupRequest {
-	int id;
-
+///  Deletes an answer bot.
+class DeleteAnswerBot extends BotRequest {
 	/// Creates a DeleteAnswerBot.
 	DeleteAnswerBot({
-			this.id = 0, 
+			super.id, 
 	});
 
 	/// Parses a DeleteAnswerBot from a string source.
@@ -538,35 +544,15 @@ class DeleteAnswerBot extends SetupRequest {
 	String _jsonType() => "DeleteAnswerBot";
 
 	@override
-	void _readProperty(String key, JsonReader json) {
-		switch (key) {
-			case "id": {
-				id = json.expectInt();
-				break;
-			}
-			default: super._readProperty(key, json);
-		}
-	}
-
-	@override
-	void _writeProperties(JsonSink json) {
-		super._writeProperties(json);
-
-		json.addKey("id");
-		json.addNumber(id);
-	}
-
-	@override
-	R visitSetupRequest<R, A>(SetupRequestVisitor<R, A> v, A arg) => v.visitDeleteAnswerBot(this, arg);
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg) => v.visitDeleteAnswerBot(this, arg);
 
 }
 
-class CheckAnswerBot extends SetupRequest {
-	int id;
-
+///  Checks whether an anwer bot has successfully registered to its Fritz!Box.
+class CheckAnswerBot extends BotRequest {
 	/// Creates a CheckAnswerBot.
 	CheckAnswerBot({
-			this.id = 0, 
+			super.id, 
 	});
 
 	/// Parses a CheckAnswerBot from a string source.
@@ -585,10 +571,92 @@ class CheckAnswerBot extends SetupRequest {
 	String _jsonType() => "CheckAnswerBot";
 
 	@override
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg) => v.visitCheckAnswerBot(this, arg);
+
+}
+
+///  Retrieves a list of calls this answer bot has already answered.
+class ListCalls extends BotRequest {
+	/// Creates a ListCalls.
+	ListCalls({
+			super.id, 
+	});
+
+	/// Parses a ListCalls from a string source.
+	static ListCalls? fromString(String source) {
+		return read(JsonReader.fromString(source));
+	}
+
+	/// Reads a ListCalls instance from the given reader.
+	static ListCalls read(JsonReader json) {
+		ListCalls result = ListCalls();
+		result._readContent(json);
+		return result;
+	}
+
+	@override
+	String _jsonType() => "ListCalls";
+
+	@override
+	R visitBotRequest<R, A>(BotRequestVisitor<R, A> v, A arg) => v.visitListCalls(this, arg);
+
+}
+
+///  Answer to a {@link ListCalls} request.
+class ListCallsResponse extends _JsonObject {
+	///  The total number of calls answered so far.
+	int callsAnswered;
+
+	///  The total amout of time taked to SPAM callers so far.
+	int talkTime;
+
+	///  The last calls that have been answered.
+	List<CallInfo> calls;
+
+	/// Creates a ListCallsResponse.
+	ListCallsResponse({
+			this.callsAnswered = 0, 
+			this.talkTime = 0, 
+			this.calls = const [], 
+	});
+
+	/// Parses a ListCallsResponse from a string source.
+	static ListCallsResponse? fromString(String source) {
+		return read(JsonReader.fromString(source));
+	}
+
+	/// Reads a ListCallsResponse instance from the given reader.
+	static ListCallsResponse read(JsonReader json) {
+		ListCallsResponse result = ListCallsResponse();
+		result._readContent(json);
+		return result;
+	}
+
+	@override
+	String _jsonType() => "ListCallsResponse";
+
+	@override
 	void _readProperty(String key, JsonReader json) {
 		switch (key) {
-			case "id": {
-				id = json.expectInt();
+			case "callsAnswered": {
+				callsAnswered = json.expectInt();
+				break;
+			}
+			case "talkTime": {
+				talkTime = json.expectInt();
+				break;
+			}
+			case "calls": {
+				json.expectArray();
+				calls = [];
+				while (json.hasNext()) {
+					if (!json.tryNull()) {
+						var value = CallInfo.read(json);
+						if (value != null) {
+							calls.add(value);
+						}
+					}
+				}
 				break;
 			}
 			default: super._readProperty(key, json);
@@ -599,12 +667,87 @@ class CheckAnswerBot extends SetupRequest {
 	void _writeProperties(JsonSink json) {
 		super._writeProperties(json);
 
-		json.addKey("id");
-		json.addNumber(id);
+		json.addKey("callsAnswered");
+		json.addNumber(callsAnswered);
+
+		json.addKey("talkTime");
+		json.addNumber(talkTime);
+
+		json.addKey("calls");
+		json.startArray();
+		for (var _element in calls) {
+			_element.writeContent(json);
+		}
+		json.endArray();
+	}
+
+}
+
+///  Information about a SPAM call answered.
+class CallInfo extends _JsonObject {
+	///  The phone number of the caller.
+	String caller;
+
+	///  The time the call has started.
+	int started;
+
+	///  The duration of the call in milliseconds.
+	int duration;
+
+	/// Creates a CallInfo.
+	CallInfo({
+			this.caller = "", 
+			this.started = 0, 
+			this.duration = 0, 
+	});
+
+	/// Parses a CallInfo from a string source.
+	static CallInfo? fromString(String source) {
+		return read(JsonReader.fromString(source));
+	}
+
+	/// Reads a CallInfo instance from the given reader.
+	static CallInfo read(JsonReader json) {
+		CallInfo result = CallInfo();
+		result._readContent(json);
+		return result;
 	}
 
 	@override
-	R visitSetupRequest<R, A>(SetupRequestVisitor<R, A> v, A arg) => v.visitCheckAnswerBot(this, arg);
+	String _jsonType() => "CallInfo";
+
+	@override
+	void _readProperty(String key, JsonReader json) {
+		switch (key) {
+			case "caller": {
+				caller = json.expectString();
+				break;
+			}
+			case "started": {
+				started = json.expectInt();
+				break;
+			}
+			case "duration": {
+				duration = json.expectInt();
+				break;
+			}
+			default: super._readProperty(key, json);
+		}
+	}
+
+	@override
+	void _writeProperties(JsonSink json) {
+		super._writeProperties(json);
+
+		json.addKey("caller");
+		json.addString(caller);
+
+		json.addKey("started");
+		json.addNumber(started);
+
+		json.addKey("duration");
+		json.addNumber(duration);
+	}
 
 }
 
@@ -627,6 +770,9 @@ class AnswerbotInfo extends _JsonObject {
 
 	///  The total number of calls accepted by this bot so far.
 	int callsAccepted;
+
+	///  The total time in milliseconds taked to SPAM customers.
+	int talkTime;
 
 	///  The name of the box to register at.
 	String registrar;
@@ -663,6 +809,7 @@ class AnswerbotInfo extends _JsonObject {
 				this.registered = false, 
 				this.registerMsg, 
 				this.callsAccepted = 0, 
+				this.talkTime = 0, 
 				this.registrar = "", 
 				this.realm = "", 
 				this.userName = "", 
@@ -714,6 +861,10 @@ class AnswerbotInfo extends _JsonObject {
 				}
 				case "callsAccepted": {
 					callsAccepted = json.expectInt();
+					break;
+				}
+				case "talkTime": {
+					talkTime = json.expectInt();
 					break;
 				}
 				case "registrar": {
@@ -780,6 +931,9 @@ class AnswerbotInfo extends _JsonObject {
 
 			json.addKey("callsAccepted");
 			json.addNumber(callsAccepted);
+
+			json.addKey("talkTime");
+			json.addNumber(talkTime);
 
 			json.addKey("registrar");
 			json.addString(registrar);
