+++ phoneblock_answerbot_ui/lib/switchIcon.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+final MaterialStateProperty<Icon?> switchIcon = MaterialStateProperty.resolveWith<Icon?>(
+      (Set<MaterialState> states) {
+    if (states.contains(MaterialState.selected)) {
+      return const Icon(Icons.check);
+    }
+    return const Icon(Icons.close);
+  },
+);
